fix : 로그인 여부에 대한 QnA Board 글 작성, 수정, 삭제 권한 제한 기능 구현

+++ qnaboard/views.py
@@ -1,6 +1,6 @@
-from django.http import HttpResponse
-from django.shortcuts import render, redirect, get_object_or_404
-from requests import post
+from django.core.exceptions import PermissionDenied
+
+from django.shortcuts import render, redirect
 
 from qnaboard.forms import PostForm
 from qnaboard.models import Post
@@ -22,18 +22,21 @@ def qb_read(request, post_id):
 
 # QnA 게시글 작성
 def qb_create(request):
-    if request.method == "POST":
-        post_form = PostForm(request.POST)
+    if not request.user.is_authenticated:
+        return redirect("qnaboard:list")
 
-        if post_form.is_valid():
-            new_post = post_form.save(commit=False)
-            new_post.save()
-            return redirect("qnaboard:list")
+    if request.user.is_authenticated:
+        if request.method == "POST":
+            post_form = PostForm(request.POST, user=request.user)
 
-    else:
-        post_form = PostForm()
+            if post_form.is_valid():
+                post_form.save()
+                return redirect("qnaboard:list")
 
-    return render(request, "qnaboard/qna_create.html", {"post_form": post_form})
+        else:
+            post_form = PostForm(user=request.user)
+
+        return render(request, "qnaboard/qna_create.html", {"post_form": post_form})
 
 
 # QnA 게시글 수정
@@ -41,21 +44,24 @@ def qb_update(request, post_id):
     post = Post.objects.get(id=post_id)
 
     if request.method == "POST":
-        post.qb_title = request.POST["qb_title"]
-        post.qb_content = request.POST["qb_content"]
+        post_form = PostForm(request.POST, instance=post, user=request.user)
 
-        post.save()
-        # return redirect("qnaboard:list")  # 리스트 페이지로 이동
-        return redirect("qnaboard:read", post_id=post.id)  # 수정된 게시물의 상세 페이지로 이동
+        if post_form.is_valid():
+            post_form.save()
+            return redirect("qnaboard:read", post_id=post.id)
 
     else:
-        updateForm = PostForm(instance=post)
-        return render(request, "qnaboard/qna_update.html", {"updateForm": updateForm})
+        post_form = PostForm(instance=post, user=request.user)
+
+    return render(request, "qnaboard/qna_update.html", {"updateForm": post_form})
 
 
 # QnA 게시글 삭제
 def qb_delete(request, post_id):
     post = Post.objects.get(id=post_id)
+    if request.user != post.user:  # 작성자만 삭제 가능
+        raise PermissionDenied  # 권한 없음
+
     post.delete()
     return redirect("qnaboard:list")
